fixed: a bug where ai_score can be a float but RetrieveResponse was only accepting int

+++ models/RetrieveResponse.py
@@ -7,4 +7,4 @@ from models.SearchResult import SearchResult
 class RetrieveResponse(BaseModel):
     results: List[SearchResult]
     answer: Dict[str,Any] = Field(..., description="The generated answer from the LLM")
-    ai_score: int = Field(..., description="A score out of 100 to detect whether content is AI written")
+    ai_score: float = Field(..., description="A score out of 100 to detect whether content is AI written")
